// src/api/generated/model/getProductBySlug200SimpleProduct.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Nero API
 * A Nero API é uma interface RESTful robusta projetada especificamente para o varejo de moda online. Ela oferece controle total sobre o ciclo de vida do produto, desde a entrada de estoque em múltiplas variantes (SKUs) até o processamento final do checkout. Ideal para marketplaces, lojas boutique ou aplicativos móveis de moda.
 * OpenAPI spec version: 1.0.0
 */

export type GetProductBySlug200SimpleProduct = {
  price: {
  current: {
  cents: number;
  value: number;
  formatted: string;
};
  original: {
  cents: number;
  value: number;
  formatted: string;
} | null;
  discountPercent: number | null;
};
  stock: number;
  skuCode: string;
  ean: string | null;
  isOutOfStock: boolean;
} | null;